update reference assets test to test with all plugins
