+++ plugins/reference-assets/core/src/plugin.ts
@@ -6,21 +6,21 @@ import {
   AsyncNodePluginPlugin,
 } from "@player-ui/async-node-plugin";
 
-const metaPlugin = new MetaPlugin([
-  new AsyncNodePlugin({
-    plugins: [new AsyncNodePluginPlugin()],
-  }),
-  new ReferenceAssetsTransformPlugin(),
-  new ChatUiDemoPlugin(),
-]);
-
 /**
  * A plugin to add transforms for the reference assets
  */
 export class ReferenceAssetsPlugin implements PlayerPlugin {
   name = "reference-assets-plugin";
 
+  private readonly metaPlugin = new MetaPlugin([
+    new AsyncNodePlugin({
+      plugins: [new AsyncNodePluginPlugin()],
+    }),
+    new ReferenceAssetsTransformPlugin(),
+    new ChatUiDemoPlugin(),
+  ]);
+
   apply(player: Player): void {
-    player.registerPlugin(metaPlugin);
+    player.registerPlugin(this.metaPlugin);
   }
 }
